requirements.txt
numpy==1.23.4
python-magic
python-magic
sympy==1.13.1
langchain
langchain-unstructured
langchain-openai
langchain_community
unstructured
unstructured-inference
streamlit
python-dotenv
pinecone-client
tiktoken
pi-heif
unstructured[image]
python-magic
cmake
onnx==1.16.1
onnxruntime　　
